:tada: lint JSON and write TS tests
