+++ .eslintrc.json
@@ -1,29 +1,47 @@
 {
-    "root": true,
-    "parser": "@typescript-eslint/parser",
-    "parserOptions": {
-        "ecmaVersion": 6,
-        "sourceType": "module",
-        "project": ["./client/tsconfig.json", "./server/tsconfig.json"]
-    },
-    "plugins": [
-        "json-format",
-        "@typescript-eslint"
+  "root": true,
+  "parser": "@typescript-eslint/parser",
+  "parserOptions": {
+    "ecmaVersion": 6,
+    "sourceType": "module",
+    "project": [
+      "./cypress/tsconfig.json",
+      "./client/tsconfig.json",
+      "./server/tsconfig.json"
+    ]
+  },
+  "plugins": [
+    "json-format",
+    "@typescript-eslint"
+  ],
+  "extends": "standard-with-typescript",
+  "rules": {
+    "@typescript-eslint/space-before-function-paren": [
+      "error",
+      {
+        "anonymous": "always",
+        "named": "never",
+        "asyncArrow": "always"
+      }
     ],
-    "extends": "standard-with-typescript",
-    "rules": {
-        "@typescript-eslint/space-before-function-paren": ["error", {
-            "anonymous": "always",
-            "named": "never",
-            "asyncArrow": "always"
-        }],
-        "space-before-function-paren": ["error", {
-            "anonymous": "always",
-            "named": "never",
-            "asyncArrow": "always"
-        }],
-        "no-void": ["error", {
-            "allowAsStatement": true
-        }]
-    }
+    "space-before-function-paren": [
+      "error",
+      {
+        "anonymous": "always",
+        "named": "never",
+        "asyncArrow": "always"
+      }
+    ],
+    "no-void": [
+      "error",
+      {
+        "allowAsStatement": true
+      }
+    ]
+  },
+  "settings": {
+    "json/json-with-comments-files": [
+      ".vscode/**"
+    ]
+  }
 }
