+++ cypress/index.d.ts
@@ -1,0 +1,8 @@
+declare namespace Cypress {
+  interface Chainable {
+    /**
+     * Select an element to DnD. See ./support/ for details.
+     */
+    dnd: (selector: string, options?: {offsetX?: number, offsetY?: number}) => Chainable<Element>
+  }
+}
