+++ cypress/plugins/index.ts
@@ -1,17 +1,17 @@
-/// <reference types="cypress" />
 // ***********************************************************
-// This example plugins/index.js can be used to load plugins
-// https://on.cypress.io/plugins-guide
+// See https://on.cypress.io/plugins-guide
 // This function is called when a project is opened or re-opened (e.g. due to
 // the project's config changing)
+
+import * as CoverageTask from '@cypress/code-coverage/task'
 
 /**
  * @type {Cypress.PluginConfig}
  */
-module.exports = (on, config) => {
+export = (on, config) => {
   // `on` is used to hook into various events Cypress emits
   // `config` is the resolved Cypress config
 
-  require('@cypress/code-coverage/task')(on, config)
+  CoverageTask(on, config)
   return config
 }
