--- cypress/support/index.js
@@ -1,1 +1,0 @@
-import '@cypress/code-coverage/support'
