+++ package.json
@@ -20,7 +20,7 @@
 		"webpack:watch": "npm run _webpack_ -- --mode development --watch",
 		"package": "vsce package",
 		"watch": "tsc -watch --build",
-		"lint": "eslint **/package.json ./.eslintrc.json **/webpack.config.js ./cypress.json ./cypress/**/*.*s ./client/src/ ./server/src/ --ext ts,js,jsx,tsx,json",
+		"lint": "eslint **/package.json ./tsconfig*.json **/tsconfig*.json ./.eslintrc.json **/webpack.config.js ./cypress.json ./cypress/ ./client/src/ ./server/src/ --ext ts,js,jsx,tsx,json",
 		"lint:fix": "npm run lint -- --fix",
 		"test:compile": "./scripts/pretest.sh",
 		"test:client": "node ./client/out/test/runTest.js",
