+++ server/package.json
@@ -1,17 +1,17 @@
 {
-	"name": "server",
-	"devDependencies": {
-		"@types/mocha": "^8.2.1",
-		"@types/mock-fs": "^4.13.0",
-		"@types/xmldom": "^0.1.30",
-		"mocha": "^8.3.0",
-		"mock-fs": "^4.13.0"
-	},
-	"dependencies": {
-		"vscode-languageserver": "^7.0.0",
-		"vscode-languageserver-textdocument": "^1.0.1",
-		"vscode-uri": "^3.0.2",
-		"xmldom": "^0.4.0",
-		"xpath-ts": "^1.3.13"
-	}
+  "name": "server",
+  "dependencies": {
+    "vscode-languageserver": "^7.0.0",
+    "vscode-languageserver-textdocument": "^1.0.1",
+    "vscode-uri": "^3.0.2",
+    "xmldom": "^0.4.0",
+    "xpath-ts": "^1.3.13"
+  },
+  "devDependencies": {
+    "@types/mocha": "^8.2.1",
+    "@types/mock-fs": "^4.13.0",
+    "@types/xmldom": "^0.1.30",
+    "mocha": "^8.3.0",
+    "mock-fs": "^4.13.0"
+  }
 }
