+++ tsconfig.base.json
@@ -1,17 +1,17 @@
 {
-	"compilerOptions": {
-		"composite": true,                     /* Enable project compilation */
-		"target": "ES2019",                    /* Specify ECMAScript target version: 'ES3' (default), 'ES5', 'ES2015', 'ES2016', 'ES2017', 'ES2018', 'ES2019' or 'ESNEXT'. */
-		"module": "commonjs",                  /* Specify module code generation: 'none', 'commonjs', 'amd', 'system', 'umd', 'es2015', or 'ESNext'. */
-		"strict": true,                        /* Enable all strict type-checking options. */
-		"declaration": true,                   /* Generates corresponding '.d.ts' file. */
-		"declarationMap": true,                /* Generates a sourcemap for each corresponding '.d.ts' file. */
-		"sourceMap": true,                     /* Generates corresponding '.map' file. */
-		"forceConsistentCasingInFileNames": true,
-		"experimentalDecorators": true,
-		"esModuleInterop": true
-	},
-	"exclude": [
-		"out"
-	]
+  "compilerOptions": {
+    "composite": true,
+    "target": "ES2019",
+    "module": "commonjs",
+    "strict": true,
+    "declaration": true,
+    "declarationMap": true,
+    "sourceMap": true,
+    "forceConsistentCasingInFileNames": true,
+    "experimentalDecorators": true,
+    "esModuleInterop": true
+  },
+  "exclude": [
+    "out"
+  ]
 }
